+++ packages/plugin-packages/mashroom-http-proxy/type-definitions/index.js
@@ -21,31 +21,49 @@ export interface MashroomHttpProxyService {
 
 }
 
-export type MashroomHttpProxyInterceptorResult = {
+export type MashroomHttpProxyRequestInterceptorResult = {
     addHeaders?: HttpHeaders;
     removeHeaders?: Array<string>;
     addQueryParams?: QueryParams;
     removeQueryParams?: Array<string>;
     rewrittenTargetUri?: string;
-    reject?: boolean;
-    rejectStatusCode?: number;
-    rejectReason?: string;
+    responseHandled?: boolean;
+}
+
+export type MashroomHttpProxyResponseInterceptorResult = {
+    addHeaders?: HttpHeaders;
+    removeHeaders?: Array<string>;
+    responseHandled?: boolean;
 }
 
 interface MashroomHttpProxyInterceptor {
 
     /**
-     * Intercept HTTP proxy call to given targetUri.
+     * Intercept request to given targetUri.
      *
      * The existingHeaders contain the original request headers, headers added by the MashroomHttpProxyService client and the ones already added by other interceptors.
      * The existingQueryParams contain query parameters from the request and the ones already added by other interceptors.
      *
-     * req is the request that shall be forwarded. DO NOT MANIPULATE IT. Just use it to access req.method and req.pluginContext.
+     * clientRequest is the request that shall be forwarded. DO NOT MANIPULATE IT. Just use it to access "method" and "pluginContext".
      *
      * Return null or undefined if you don't want to interfere with a call.
      */
-    intercept(targetUri: string, existingHeaders: HttpHeaders, existingQueryParams: QueryParams, req: ExpressRequest): Promise<?MashroomHttpProxyInterceptorResult>;
+    interceptRequest(targetUri: string, existingHeaders: HttpHeaders, existingQueryParams: QueryParams,
+                     clientRequest: ExpressRequest, clientResponse: ExpressResponse):
+        Promise<?MashroomHttpProxyRequestInterceptorResult>;
+
+    /**
+     * Intercept response from given targetUri.
+     *
+     * The existingHeaders contain the original request header and the ones already added by other interceptors.
+     * targetResponse is the response that shall be forwarded to the client. DO NOT MANIPULATE IT. Just use it to access "statusCode" an such.
+     *
+     * Return null or undefined if you don't want to interfere with a call.
+     */
+    interceptResponse(targetUri: string, existingHeaders: HttpHeaders, targetResponse: http$IncomingMessage<>, clientRequest: ExpressRequest, clientResponse: ExpressResponse):
+        Promise<?MashroomHttpProxyResponseInterceptorResult>;
 }
+
 
 /*
  * Bootstrap method definition for http-proxy-interceptor plugins
