+++ packages/plugin-packages/mashroom-http-proxy/type-definitions/internal.ts
@@ -1,6 +1,7 @@
 
-import type {IncomingHttpHeaders} from 'http';
-import type {MashroomHttpProxyInterceptor} from './api';
+import type {IncomingHttpHeaders, IncomingMessage} from 'http';
+import type {ExpressRequest, ExpressResponse, MashroomLogger} from '@mashroom/mashroom/type-definitions';
+import type {MashroomHttpProxyInterceptor, MashroomHttpProxyResponseInterceptorResult, HttpHeaders, MashroomHttpProxyRequestInterceptorResult} from './api';
 
 export interface HttpHeaderFilter {
     filter(headers: IncomingHttpHeaders): void;
@@ -26,4 +27,17 @@ export interface MashroomHttpProxyInterceptorRegistry {
     readonly interceptors: Array<MashroomHttpProxyInterceptorHolder>;
     register(pluginName: string, interceptor: MashroomHttpProxyInterceptor): void;
     unregister(pluginName: string): void;
+}
+
+export interface InterceptorHandler {
+    processRequest(clientRequest: ExpressRequest, clientResponse: ExpressResponse,
+                   targetUri: string, additionalHeaders: HttpHeaders, logger: MashroomLogger):
+        Promise<MashroomHttpProxyRequestInterceptorResult>;
+    processResponse(clientRequest: ExpressRequest, clientResponse: ExpressResponse, targetUri: string,
+                    targetResponse: IncomingMessage, logger: MashroomLogger):
+        Promise<MashroomHttpProxyResponseInterceptorResult>;
+}
+
+export interface Proxy {
+    forward(req: ExpressRequest, res: ExpressResponse, uri: string, additionalHeaders: HttpHeaders): Promise<void>
 }
